requirements.txt
requests
python-telegram-bot
APScheduler
python-dotenv
pytz
redis